+++ CHANGELOG.md
@@ -10,7 +10,8 @@ phpSysInfo 3.3.3
  - [ADD] Croatian Translation hr.xml
  - [ADD] Milis, FreeNAS and UOS to detected distros
  - [ADD] SMART plugin - reading contents of wmic command for WinNT
-
+ - [ADD] Raid plugin - 3ware-status status
+ 
  - [NEW] Ability to read information from an external WinNT server (new WMI_HOSTNAME, WMI_USER and WMI_PASSWORD parameters)
  - [NEW] NvidiaSMI - sensors monitoring for Nvidia GPU
  - [NEW] plugin Viewer - show output of any command or file viewer.txt contents